Improve *for and add redraw()
redraw() is added to redraw all bindings in all components. It's there
to help when changes occur deeply in objects and immutable objects are
not used.

*for directive had issues because it would move elements in the DOM.
This would disconnect and reconnect the controller while the element was
removed and then added back. This is bad because properties can not be
restored from the element into the controller.

*for is a place that could use some serious attention for performance
gains.