1.8.0 - Fix conflicting attr and prop
Consider an element with both attr and prop defined.

Previously, Fudgel would take these steps.

    * During conectedCallback, synchronously bind
    * For each prop
        * Set up a setter to detect changes
        * Set the controller to the current property value
    * For each attr
        * Set the controller to the current attr value
        * Watch for updates

This means attributes would beat properties. This is bad because
JavaScript could create attributes before properties or vice-versa. The
attributes and properties are all set before the element is added to the
document, so now there needs to be fallbacks and priorities.

    * During connectedCallback, synchronously bind
    * For each attr
        * Set the controller to the current attr value ONLY IF IT EXISTS
        * Watch for updates
    * For each prop
        * Set up the controller to the current property value ONLY IF IT
          EXISTS (even if it is undefined)
        * Set up a setter to detect changes

What we really want is consistency. When instantiated, the custom
element will set up bindings. The bindings will first scan for
attributes, setting up listeners and setting controller values from
attributes. Next, the bindings will set up property change handlers for
properties. It will no longer set an initial value from a property.

The order of priorities. The higher items on the list are overridden by
the lower items.

1. Initial class value.
2. Attribute value, when the attribute exists on the element during
   connectedCallback().
3. Property value (even `undefined`) when the property exists on the
   element instance during connectedCallback().
4. Attribute and property changes after connectedCallback(), with newer
   changes overriding older changes.