Handle attribute and property bindings better
When elements were removed and added back to the DOM, attributes could
get initialized twice or have their event handlers added multiple times.
The controller instance that was used may have been the wrong one. Same
goes for properties, but their problems were easier to detect - they
simply didn't work right away.

Fixed how the weakmaps are used in metadata - if a second argument is
passed, it is an unconditional assignment, which fixed the update to the
active controller instance that is assigned to an element.

Fixed multiple patchers being added to a single patched setter by
removing the callbacks when the element is disconnected - reconnecting
will reinitialize the controller and reassign the callbacks.

Fixed <slot-like> and made it restore the DOM to what it looked like
before processing, which will better simulate what happens with shadow
DOMs. This means removing and adding a light dom that uses <slot-like>
will now avoid duplication of content.